// SDKs/twilio.com/twilio_verify_v2/1.37.3/typescript/src/sdk/models/shared/verifyv2servicemessagingconfiguration.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class VerifyV2ServiceMessagingConfiguration extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=country" })
  country?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=messaging_service_sid" })
  messagingServiceSid?: string;

  @SpeakeasyMetadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}